pyproject.toml
[project]
name = "bacpipe"
version = "1.1"
description = "Use bacpipe to streamline the process of generating embeddings and analysing your PAM datasets."
readme = "README.md"
authors = [
    { name = "Vincent S. Kather", email = "vkather@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "absl-py==2.3.1",
    "anyio==4.10.0",
    "apache-beam==2.60.0",
    "astunparse==1.6.3",
    "attrs==25.3.0",
    "audioread==3.0.1",
    "bleach==6.2.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "chex==0.1.90",
    "click==8.2.1",
    "cloudpickle==2.2.1",
    "contextlib2==21.6.0",
    "contourpy==1.3.3",
    "crcmod==1.7",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "dill==0.3.1.1",
    "dnspython==2.7.0",
    "docopt==0.6.2",
    "einops==0.8.0",
    "etils==1.13.0",
    "fastavro==1.12.0",
    "fasteners==0.20",
    "filelock==3.19.1",
    "flatbuffers==25.2.10",
    "flax==0.10.0",
    "fonttools==4.59.2",
    "fsspec==2025.9.0",
    "gast==0.6.0",
    "google-auth==2.40.3",
    "google-auth-oauthlib==1.2.2",
    "google-pasta==0.2.0",
    "grpcio==1.65.5",
    "h11==0.16.0",
    "h5py==3.14.0",
    "hdfs==2.7.3",
    "hear21passt==0.0.26",
    "hf-xet==1.1.9",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.27.0",
    "huggingface-hub==0.34.4",
    "humanize==4.13.0",
    "idna==3.10",
    "importlib-resources==6.5.2",
    "iniconfig==2.1.0",
    "jax==0.4.33",
    "jaxlib==0.4.33",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonpickle==3.3.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.4.1",
    "keras==2.15.0",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "libclang==18.1.1",
    "librosa==0.10.2.post1",
    "linkify-it-py==2.0.3",
    "llvmlite==0.44.0",
    "markdown==3.9",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.6",
    "mdit-py-plugins==0.5.0",
    "mdurl==0.1.2",
    "ml-collections==0.1.0",
    "ml-dtypes==0.3.2",
    "mpmath==1.3.0",
    "msgpack==1.1.1",
    "narwhals==2.3.0",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "numba==0.61.2",
    "numpy==1.26.4",
    "oauthlib==3.3.1",
    "objsize==0.7.1",
    "opt-einsum==3.4.0",
    "optax==0.2.3",
    "orbax-checkpoint==0.6.4",
    "orjson==3.10.7",
    "packaging==25.0",
    "pandas==2.2.3",
    "panel==1.6.2",
    "param==2.2.1",
    "pillow==11.3.0",
    "platformdirs==4.4.0",
    "pluggy==1.6.0",
    "pooch==1.8.2",
    "proto-plus==1.26.1",
    "protobuf==4.25.8",
    "pyarrow==16.1.0",
    "pyarrow-hotfix==0.7",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydot==1.4.2",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pymongo==4.14.1",
    "pynndescent==0.5.13",
    "pyparsing==3.2.3",
    "pytest==8.4.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "redis==5.3.1",
    "referencing==0.36.2",
    "regex==2025.9.1",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rich==14.1.0",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "safetensors==0.4.5",
    "scikit-learn==1.7.1",
    "scipy==1.13.1",
    "seaborn==0.13.2",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sounddevice==0.4.7",
    "soundfile==0.12.1",
    "soxr==0.3.7",
    "sympy==1.13.1",
    "tensorboard==2.15.2",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.15.1",
    "tensorflow-estimator==2.15.0",
    "tensorflow-hub==0.16.0",
    
    # Packages to skip on Windows
    "bokeh==3.7.3; sys_platform != 'win32'",
    "nvidia-cublas-cu12==12.4.5.8; sys_platform != 'win32'",
    "nvidia-cuda-cupti-cu12==12.4.127; sys_platform != 'win32'",
    "nvidia-cuda-nvrtc-cu12==12.4.127; sys_platform != 'win32'",
    "nvidia-cuda-runtime-cu12==12.4.127; sys_platform != 'win32'",
    "nvidia-cudnn-cu12==9.1.0.70; sys_platform != 'win32'",
    "nvidia-cufft-cu12==11.2.1.3; sys_platform != 'win32'",
    "nvidia-curand-cu12==10.3.5.147; sys_platform != 'win32'",
    "nvidia-cusolver-cu12==11.6.1.9; sys_platform != 'win32'",
    "nvidia-cusparse-cu12==12.3.1.170; sys_platform != 'win32'",
    "nvidia-cusparselt-cu12==0.6.2; sys_platform != 'win32'",
    "nvidia-nccl-cu12==2.21.5; sys_platform != 'win32'",
    "nvidia-nvjitlink-cu12==12.4.127; sys_platform != 'win32'",
    "nvidia-nvtx-cu12==12.4.127; sys_platform != 'win32'",
    "triton==3.2.0; sys_platform != 'win32'",
    "uvloop==0.20.0; sys_platform != 'win32'",

    "tensorstore==0.1.74",
    "termcolor==3.1.0",
    "threadpoolctl==3.6.0",
    "timm==1.0.11",
    "tokenizers==0.22.0",
    "toolz==1.0.0",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "transformers==4.56.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "uc-micro-py==1.0.3",
    "umap-learn==0.5.9.post2",
    "urllib3==2.5.0",
    "uvicorn==0.23.2",
    "webencodings==0.5.1",
    "werkzeug==3.1.3",
    "wheel==0.45.1",
    "wrapt==1.14.2",
    "xyzservices==2025.4.0",
    "zipp==3.23.0",
    "zstandard==0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.package-data]
"bacpipe" = ["*.yaml", "*.png"]